Daemonize the scheduler process unless SCCACHE_NO_DAEMON is set in the environment. Fixes #382
